Add missing open message tests
